chore: Moved chat styles to single file and separated MessageBubble component
