--- ui/src/components/features/chat/Chat/chat.module.css
@@ -1,51 +1,0 @@
-.container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.chatBody {
-  overflow-y: auto;
-  outline: 1px solid #484848;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  max-height: 100%;
-  gap: 1rem;
-  flex-grow: 1;
-  @apply border border-muted rounded;
-}
-
-.chatItem {
-  @apply p-2 px-4 rounded bg-theme-surface0 flex flex-col gap-1 w-fit;
-}
-
-.chatItemRight {
-  @apply ml-auto;
-}
-
-.chatItemHeading {
-  @apply text-theme-blue;
-  @apply p-2 rounded bg-theme-surface0 flex flex-col gap-1 w-full;
-}
-
-.chatItemHeading {
-  @apply text-theme-blue;
-}
-
-.status {
-  padding: 0.5rem;
-  border-radius: 50px;
-  text-align: center;
-  display: inline-block;
-}
-
-.status.connected {
-  @apply bg-green-600 text-background;
-}
-
-.status.disconnected {
-  @apply bg-red-600 text-base;
-}
